Bump dash-functional version and deprecate it
This is the last release of dash-functional.el before it is absorbed
by dash.el.  The file dash-functional.el will no longer be
maintained, but it will continue to be provided on the branch
compat/dash-functional for backward-compatibility with the separate
dash-functional package distributed on MELPA.  Users are encouraged
to switch to dash 2.18.0 instead when that becomes available.  For
discussion, see issue #356.

* dash-functional.el: Update copyright notice for 2021.  Bump
version to 1.3.0.  Mention deprecation in package commentary.
* dash-template.texi:
* readme-template.md: Bump dash-functional version mentioned.
* README.md:
* dash.texi: Regenerate docs.